Aplit call and state into own files Change to `publish_state` Use LogString Force implementation of virtual functions
